backend/requirements.txt
flask
flask-cors
lyricsgenius
deep-translator
google-generativeai
requests
python-dotenv
langdetect
